requirements.txt
pdfplumber
python-docx
requests
sentence-transformers
faiss-cpu
google-generativeai